Reorganize the source code
Entry point to the program moved to cmd directory while all
supplementary files and unit tests now stored in pkg/utils path as
sub-package. This reorganization makes navigation through the project
more convenient
