+++ pkg/utils/data.go
@@ -1,0 +1,20 @@
+package utils
+
+import "time"
+
+type AgentInfo struct {
+	ReportInterval int                 `json:"report_interval"`
+	PodName        string              `json:"podname"`
+	HostDate       time.Time           `json:"hostdate"`
+	LastUpdated    time.Time           `json:"last_updated"`
+	LookupHost     map[string][]string `json:"nslookup"`
+	IPs            map[string][]string `json:"ips"`
+}
+
+type CheckConnectivityInfo struct {
+	Message  string   `json="message"`
+	Absent   []string `json="outdated,omitempty"`
+	Outdated []string `json="absent,omitempty"`
+}
+
+var AgentCache = make(map[string]AgentInfo)
